- Fixed addToPot not working - Client no longer handles prevBet - Other fixes/stuff

+++ src/Server/src/Utils/Database/Database.java
@@ -80,7 +80,7 @@ public class Database {
                 .executeUpdate();
     }
 
-    public void updateWallet(int userID, int amount) {
+    public void updateWallet(int userID, double amount) {
         String sql = "UPDATE Players p SET p.wallet = :wallet WHERE p.user_id = :userID";
         Connection con = sql2o.open();
         con.createQuery(sql)
